+++ src/app/app.module.ts
@@ -1,9 +1,12 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 
 import { AckFxModule } from 'ack-angular-fx';
 import { AckModule } from 'ack-angular/AckModule';
+import { AckOfflineModule } from 'ack-angular/modules/offline/module';
+
 import { AppRoutingModule } from './app-routing.module';
 import { AckRouterModule } from "ack-angular/AckRouterModule"
 import { NgxPageScrollModule } from "ngx-page-scroll"
@@ -24,8 +27,10 @@ import { GotoTop } from './components/goto-top/GotoTop.directive';
     BrowserAnimationsModule,
     NgxPageScrollModule,
     AckRouterModule,
+    FormsModule,
     AppRoutingModule,
     AckModule.forRoot(),
+    AckOfflineModule.forRoot(),
     AckFxModule
   ],
   providers: [