+++ src/app/components/demos/Demos.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+import { string } from "./demos.template"
+import { animations } from "ack-angular-fx"
+
+@Component({
+  template: string,
+  animations:animations
+})
+export class Demos {
+  shows:any = {}
+}