+++ src/app/components/intro-family/IntroFamily.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+import { string } from "./intro-family.template"
+import { animations } from "ack-angular-fx"
+
+@Component({
+  template: string,
+  animations:animations
+}) export class IntroFamily {
+}