+++ src/app/components/purpose/Purpose.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+import { string } from "./purpose.template"
+import { animations } from "ack-angular-fx"
+
+@Component({
+  template: string,
+  animations:animations
+}) export class Purpose {
+}