+++ src/app/declarations.ts
@@ -11,16 +11,45 @@ import { BundleStats } from './components/bundle-stats/BundleStats.component';
 import { AckMeter } from './components/ack-meter/AckMeter.component';
 import { AckMeterDisplay } from './components/ack-meter-display/AckMeterDisplay.component';
 
+import { Demos } from './components/demos/Demos.component';
+import { DeviceOrientation } from './components/device-orientation/DeviceOrientation.component';
+import { UserMediaDemo } from './components/user-media-demo/UserMediaDemo.component';
+import { DeviceLocation } from './components/device-location/DeviceLocation.component';
+import { SocialSharing } from './components/social-sharing/SocialSharing.component';
+import { ShareButton } from './components/share-button/ShareButton.directive';
+
+import { Family } from './components/family/Family.component';
+import { FamilyLogin } from './components/family-login/FamilyLogin.component';
+
+import { Intro } from './components/intro/Intro.component';
+import { IntroFamily } from './components/intro-family/IntroFamily.component';
+import { Welcome } from './components/welcome/Welcome.component';
+import { Purpose } from './components/purpose/Purpose.component';
+
 export const declarations = [
+  Demos,
   AckMeterDisplay,
   AckMeter,
 
   ResumeHighlights,
+  UserMediaDemo,
+  DeviceOrientation,
+  SocialSharing,
+  DeviceLocation,
+  ShareButton,
   BundleStats,
   Passion,
   ComingSoon,
   MainMenu,
   Life,
+
+  Family,
+  FamilyLogin,
+  
+  Intro,
+  Welcome,
+  Purpose,
+  IntroFamily,
   //GotoTop,
   Resume
 ]